requirements.txt
anyio==4.0.0
astor==0.8.1
attrdict==2.0.1
Babel==2.13.1
bce-python-sdk==0.8.92
beautifulsoup4==4.12.2
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.4
decorator==5.1.1
et-xmlfile==1.1.0
fire==0.5.0
Flask==3.0.0
flask-babel==4.0.0
fonttools==4.43.1
future==0.18.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
imageio==2.31.6
imgaug==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
networkx==3.2.1
numpy==1.26.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
paddleocr==2.7.0.3
paddlepaddle==2.5.2
pandas==2.1.2
pdf2docx==0.5.6
Pillow==10.0.1
premailer==3.10.0
protobuf==4.24.4
psutil==5.9.6
pyclipper==1.3.0.post5
pycryptodome==3.19.0
PyMuPDF==1.20.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.4.0
rarfile==4.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
termcolor==2.3.0
tifffile==2023.9.26
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
visualdl==2.5.3
Werkzeug==3.0.1
